backend/pyproject.toml
[project]
name = "focusread-backend"
version = "0.1.0"
description = "FastAPI backend for FocusRead"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "pydantic>=2.6.0",
    "python-multipart>=0.0.9",
    "sqlalchemy>=2.0.45",
    "aiosqlite>=0.22.1",
    "asyncpg>=0.31.0",
    "passlib[bcrypt]>=1.7.4",
    "pyjwt>=2.10.1",
    "pypdf>=6.5.0",
    "fpdf>=1.7.2",
    "requests>=2.32.5",
    "google-genai>=1.56.0",
    "httpx>=0.28.1",
    "pdf2image>=1.17.0",
]

[dev-dependencies]
pytest = ">=8.0.0"
httpx = ">=0.27.0"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
]
